refactor(core): rename text2sql_common → pipeline
Mechanical rename. The "_common" suffix originally meant "shared between text2sql.py (streaming) and text2sql_sync.py (SDK)" — it stops meaning anything once those two collapse into one orchestration. "pipeline" describes what's actually in the file: the building blocks of the text2sql pipeline.

All importers updated: api/core/__init__.py, api/core/schema_loader.py, api/core/text2sql.py, api/core/text2sql_sync.py, api/routes/graphs.py, queryweaver_sdk/client.py.

No behavior change. Pylint 10/10, SDK fast tests pass.
